+++ packages/components/src/Tooltip/Tooltip.css.d.ts
@@ -1,5 +1,6 @@
 export const shadowActivator: string;
 export const tooltipWrapper: string;
 export const tooltip: string;
+export const arrow: string;
 export const above: string;
 export const below: string;